requirements.txt
gymnasium>=0.29.0
numpy>=1.21.0
opencv-python>=4.5.0
pillow>=8.3.0
matplotlib>=3.5.0
pytest>=7.0.0
pytest-cov>=4.0.0
pygame>=2.1.0
PyOpenGL>=3.1.0
PyOpenGL_accelerate>=3.1.0
# pyopengl-osmesa  # for headless linux rendering
